Refactor files names
